add periodic ops to history module
