--- modules/pricefeed/expected_modules.go
@@ -1,8 +1,0 @@
-package pricefeed
-
-import "github.com/forbole/bdjuno/v2/types"
-
-type HistoryModule interface {
-	IsHistoryModuleEnabled() bool 
-	UpdatePricesHistory([]types.TokenPrice) error
-}
