+++ modules/pricefeed/module.go
@@ -17,24 +17,22 @@ var (
 
 // Module represents the module that allows to get the token prices
 type Module struct {
-	cfg           *Config
-	cdc           codec.Marshaler
-	db            *database.Db
-	historyModule HistoryModule
+	cfg *Config
+	cdc codec.Marshaler
+	db  *database.Db
 }
 
 // NewModule returns a new Module instance
-func NewModule(cfg config.Config, historyModule HistoryModule, cdc codec.Marshaler, db *database.Db) *Module {
+func NewModule(cfg config.Config, cdc codec.Marshaler, db *database.Db) *Module {
 	pricefeedCfg, err := ParseConfig(cfg.GetBytes())
 	if err != nil {
 		panic(err)
 	}
 
 	return &Module{
-		cfg:           pricefeedCfg,
-		cdc:           cdc,
-		db:            db,
-		historyModule: historyModule,
+		cfg: pricefeedCfg,
+		cdc: cdc,
+		db:  db,
 	}
 }
